Function to manage battery modified, it now works in open collector mode. Tested OK.
Function to detect charging tested too in both firmware and hardware. Care must be taken when using it, as sensors have to be powered off.

safeDisconnect() and recoverVariables() functions tested OK.
RtcMemory class "var" struct members are stored together and allingned by the struct. In order to access them, writeData and readData methods must be used. The address to be read is defined into "RTC_MEMORY_VARIABLES_START_BLOCK".