transmission state machine working. Check why it is taking so long (sometimes) to log into AP. Maybe cause it's the PC AP and it's not optimised or very fast. Try using the router as AP to the server.
To do: Make code to read pending packets (also calculating all the pending ones if "packets" is not specified in functon call, as default is 0).

CHECK: It seems that checkPowerdown is not being called properly. Better put this mechanism in RtcMemory constructor, as it has more to do with RTC memory directly than with the wake sequence itself; in the constructor is more safe not to be omitted and also clears the code.

Improve retries sequence.