Force transmission fixed, now it will save all the measurements stored in memory into flash (even if it was not full) and transmit them after this. Tested OK.
Hardware: battery disconnect circuit tested OK driven with GPIO.

to do: Improve the packets-sent system: If any unsincronization happened, the sent_pointer will only increment with the amount of packets sent, without any reference. Trouble could appear if it's wanted to read all the data from archive (from web server, for instance). In that last case, this could be avoided by not incrementing the archive_sent_packet when calling the function (not using the return value to that purpose). -> in a future, web server could send commands to the device in order to read a specific portion/piece of the archive.