Get server time tested OK.
to do:
Fix goDeepSleep() bug, when called, the argument sent is the time minus the elapsed time (t-millis()). Inside the function, the time is all handled in microseconds.
To fix this, send "t" argument in milliseconds (not 60e6 but 60e6, for example) and then multiply by 1000 to convert to microseconds, to send that value to the deep sleep command from SDK.

- Data sending schedule: It is not sent when rtc memory is full anymore, data transmission must be triggered by external wake up or by the daily hour (configurable), to say, 6:00 AM.